requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
apscheduler
pytest
